Q5. Annual Energy production file created
Weibull function created. Also file that tests Weibull function and plots the probabilities and power curves created. Could be added to the main, but I choose not to so we leave it, to decide what to do with it, however Q5 uses code of Q1 but adds an extra "for loop" to run the code across the range of wind speeds